Wrapper can fail  to report a .NET Exception
If an exception is thrown whilst in the pcall the return loses the
information due to stack unwinding.
( see: http://www.lua.org/pil/8.5.html )

Thus tbl is nil and we exit irregularly with a proper LUA fail on access
to a nil value.

A better way is to call in with xpcall.

If the xpcall fails due to a .NET exception in GetComponent
then we get the stack unwind given to us via the call to teh anonymous
error handler.